refactor: single-source config package; fix universal-middleware test on TS7
- Collapse `@universal-middleware/tsdown-config` to a single `index.ts` (was a
  `index.js` + hand-written `index.d.ts` pair with duplicated docs). Importing it
  by package name means the native config loader resolves the `.ts` and tsc reads
  its types directly, with no build step, no `.d.ts` to keep in sync, and no
  declaration emitted for it.
- universal-middleware: replace `@rollup/plugin-typescript` (which calls the
  classic TS compiler API that TS7 removed — `ts.ScriptTarget` is undefined) with
  `@rollup/plugin-swc` in the rollup test, plus a small resolver for the `.ts`
  fixtures' `.js`-suffixed imports that the typescript plugin used to remap.

Build, typecheck, biome lint, and the universal-middleware tests all pass on TS7.
