+++ packages/universal-middleware/package.json
@@ -132,7 +132,7 @@
   "devDependencies": {
     "@rollup/plugin-commonjs": "^29.0.3",
     "@rollup/plugin-node-resolve": "^16.0.3",
-    "@rollup/plugin-typescript": "^12.3.0",
+    "@rollup/plugin-swc": "^0.4.1",
     "@swc/core": "catalog:",
     "@universal-middleware/tsdown-config": "workspace:*",
     "esbuild": "^0.28.1",
